+++ packages/core/src/defaults/values/border-values.ts
@@ -1,5 +1,5 @@
 import { createValues } from "@/helpers/create-values";
-import { borderRadius, borderSpacing, borderWidth } from "../variables";
+import { borderRadius, borderWidth } from "../variables";
 
 export const borderRadiusValues: Record<string, string> = {
 	...createValues({
@@ -15,12 +15,14 @@ export const borderRadiusValues: Record<string, string> = {
 };
 
 export const borderSpacingValues: Record<string, string> = {
-	...createValues({
-		base: borderSpacing,
-		unit: "rem",
-		min: 0,
-		max: 8,
-	}),
+	auto: "auto",
+	xs: "0.125rem",
+	sm: "0.25rem",
+	md: "0.5rem",
+	lg: "1rem",
+	xl: "1.5rem",
+	"xxl": "2rem",
+	"3xl": "3rem",
 	px: "1px",
 };